+++ retail/modules/nameFix.lua
@@ -77,6 +77,7 @@ local playerFrameOCDTextureBypass
 local hidePlayerName
 local hidePetName
 local isAddonLoaded = C_AddOns.IsAddOnLoaded
+local changeUnitFrameFont
 
 function BBF.UpdateUserTargetSettings()
     hidePartyNames = BetterBlizzFramesDB.hidePartyNames
@@ -99,6 +100,7 @@ function BBF.UpdateUserTargetSettings()
     playerFrameOCDTextureBypass = BetterBlizzFramesDB.playerFrameOCDTextureBypass
     hidePlayerName = BetterBlizzFramesDB.hidePlayerName
     hidePetName = BetterBlizzFramesDB.hidePetName
+    changeUnitFrameFont = BetterBlizzFramesDB.changeUnitFrameFont
 end
 
 local function CenterPlayerName()
@@ -380,6 +382,27 @@ InitializeFontStringsForFrames()
 
 
 
+local function UpdateFontStringPosition(frame)
+    local name = frame.name or frame.Name
+    if not name or not name:GetParent() then return end
+    local point, relativeTo, relativePoint, xOffset, yOffset = name:GetPoint()
+    if point then
+        frame.bbfName:ClearAllPoints()
+        frame.bbfName:SetPoint(point, relativeTo, relativePoint, xOffset, yOffset)
+    end
+end
+
+local function UpdateAllFontStringPositions()
+    for _, frame in ipairs(frames) do
+        UpdateFontStringPosition(frame)
+    end
+end
+
+C_Timer.After(1, function()
+    if C_AddOns.IsAddOnLoaded("EasyFrames") then
+        UpdateAllFontStringPositions()
+    end
+end)
 
 
 
@@ -929,7 +952,8 @@ local function SetArenaNameUnitFrame(frame, unit, textObject)
     -- Construct the nameText based on specName and unitID settings
     if specName then
         if showSpecName and showArenaID and unitID then
-            nameText = specName .. " " .. unitID
+            local arenaNumber = string.match(unitID, "%d+")
+            nameText = specName .. " " .. (arenaNumber or "")
         elseif showSpecName then
             nameText = specName
         elseif showArenaID and unitID then
@@ -956,6 +980,11 @@ local function PlayerFrameNameChanges(frame)
         frame.bbfName:SetText("")
         return
     end
+
+    if not changeUnitFrameFont then
+        frame.bbfName:SetFont(frame.name:GetFont())
+    end
+
     if classColorTargetNames then
         ClassColorName(frame.bbfName, unit)
     end
@@ -974,6 +1003,10 @@ local function TargetFrameNameChanges(frame)
     frame.name:SetAlpha(0)
     if not frame.unit then return end
     local unit = frame.unit
+
+    if not changeUnitFrameFont then
+        frame.bbfName:SetFont(frame.name:GetFont())
+    end
 
     if targetAndFocusArenaNames and IsActiveBattlefieldArena() then
         SetArenaNameUnitFrame(frame, unit, frame.bbfName)
@@ -1020,6 +1053,9 @@ local function PetFrameNameChanges(frame)
         frame.bbfName:SetText("")
         return
     end
+    if not changeUnitFrameFont then
+        frame.bbfName:SetFont(frame.name:GetFont())
+    end
     frame.bbfName:SetText(frame.name:GetText())
     if classColorTargetNames then
         ClassColorName(frame.bbfName, unit)
@@ -1045,6 +1081,10 @@ local function FocusFrameNameChanges(frame)
         local _, class = UnitClass(unit)
         local classColor = RAID_CLASS_COLORS[class]
         frame.TargetFrameContent.TargetFrameContentMain.LevelText:SetTextColor(classColor.r, classColor.g, classColor.b)
+    end
+
+    if not changeUnitFrameFont then
+        frame.bbfName:SetFont(frame.name:GetFont())
     end
 
     if targetAndFocusArenaNames and IsActiveBattlefieldArena() then
@@ -1080,6 +1120,9 @@ local function TargetFrameToTNameChanges(frame)
     frame.name:SetAlpha(0)
     if not frame.unit then return end
     local unit = frame.unit
+    if not changeUnitFrameFont then
+        frame.bbfName:SetFont(frame.name:GetFont())
+    end
     if targetAndFocusArenaNames and IsActiveBattlefieldArena() then
         SetArenaNameUnitFrame(frame, unit, frame.bbfName)
     else
@@ -1106,6 +1149,9 @@ local function FocusFrameToTNameChanges(frame)
     frame.name:SetAlpha(0)
     if not frame.unit then return end
     local unit = frame.unit
+    if not changeUnitFrameFont then
+        frame.bbfName:SetFont(frame.name:GetFont())
+    end
     if targetAndFocusArenaNames and IsActiveBattlefieldArena() then
         SetArenaNameUnitFrame(frame, unit, frame.bbfName)
     else
@@ -1176,13 +1222,14 @@ function BBF.AllNameChanges()
                 TargetFrameToT,
                 FocusFrameToT,
             }
-        
+
             -- Iterate through each frame and reset the text color
             for _, frame in pairs(frames) do
                 if frame and frame.name then
                     local a,b,c = frame.name:GetFont()
                     frame.bbfName:SetFont(a,b,c)
-                    frame.bbfName:SetTextColor(frame.name:GetTextColor())
+                    local r, g, b, a = frame.name:GetTextColor()
+                    frame.bbfName:SetTextColor(r,g,b,1)
                     if not frame.bbfhbcHook then
                         hooksecurefunc(frame.name, "SetFont", function(self)
                             local f,s,o = self:GetFont()
@@ -1190,7 +1237,8 @@ function BBF.AllNameChanges()
                             frame.bbfName:SetFont(f,s,o)
                         end)
                         hooksecurefunc(frame.name, "SetTextColor", function(self)
-                            frame.bbfName:SetTextColor(self:GetTextColor())
+                            local r, g, b, a = self:GetTextColor()
+                            frame.bbfName:SetTextColor(r,g,b,1)
                         end)
                         frame.bbfhbcHook = true
                     end