+++ temp_cata/BetterBlizzFrames.lua
@@ -897,8 +897,9 @@ function BBF.HookTextures()
     local db = BetterBlizzFramesDB
     -- Hook UnitFrames
     -- BetterBlizzFramesDB.textureSwapUnitFrames
-    HookUnitFrameTextures()
-
+    if db.changeUnitFrameHealthbarTexture or db.changeUnitFrameManabarTexture then
+        HookUnitFrameTextures()
+    end
     -- Hook Raidframes
     -- BetterBlizzFramesDB.textureSwapRaidFrames
     if db.changeRaidFrameHealthbarTexture or db.changeRaidFrameManabarTexture then