+++ temp_era/modules/healthbarColor.lua
@@ -537,6 +537,13 @@ function BBF.BiggerHealthbars(frame, name)
 end
 
 function BBF.HookBiggerHealthbars()
+    if C_AddOns.IsAddOnLoaded("DragonflightUI") then
+        if not BBF.DFUIUnsupported then
+            print("|A:gmchat-icon-blizz:16:16|aBetter|cff00c0ffBlizz|rFrames: Bigger Healthbars is not supported with DragonflightUI")
+            BBF.DFUIUnsupported = true
+        end
+        return
+    end
     if BetterBlizzFramesDB.biggerHealthbars and not biggerHealthbarHooked then
         local playerName = PlayerFrame.bbfName
         local targetName = TargetFrame.bbfName or TargetFrameTextureFrameName