+++ temp_era/modules/nameFix.lua
@@ -75,6 +75,7 @@ local classColorLevelText
 local hidePlayerName
 local hidePetName
 local isAddonLoaded = C_AddOns.IsAddOnLoaded
+local changeUnitFrameFont
 
 function BBF.UpdateUserTargetSettings()
     hidePartyNames = BetterBlizzFramesDB.hidePartyNames
@@ -95,6 +96,7 @@ function BBF.UpdateUserTargetSettings()
     classColorLevelText = BetterBlizzFramesDB.classColorLevelText
     hidePlayerName = BetterBlizzFramesDB.hidePlayerName
     hidePetName = BetterBlizzFramesDB.hidePetName
+    changeUnitFrameFont = BetterBlizzFramesDB.changeUnitFrameFont
 end
 
 local validPartyUnits = {
@@ -275,6 +277,28 @@ end
 
 -- Run the function to initialize font strings on all specified frames
 InitializeFontStringsForFrames()
+
+local function UpdateFontStringPosition(frame)
+    local name = frame.name or frame.Name
+    if not name or not name:GetParent() then return end
+    local point, relativeTo, relativePoint, xOffset, yOffset = name:GetPoint()
+    if point then
+        frame.bbfName:ClearAllPoints()
+        frame.bbfName:SetPoint(point, relativeTo, relativePoint, xOffset, yOffset)
+    end
+end
+
+local function UpdateAllFontStringPositions()
+    for _, frame in ipairs(frames) do
+        UpdateFontStringPosition(frame)
+    end
+end
+
+C_Timer.After(1, function()
+    if C_AddOns.IsAddOnLoaded("DragonflightUI") or C_AddOns.IsAddOnLoaded("EasyFrames") then
+        UpdateAllFontStringPositions()
+    end
+end)
 
 local function SetPartyFont(font, size, outline, size2)
     if outline == "NONE" then
@@ -775,7 +799,8 @@ local function SetArenaNameUnitFrame(frame, unit, textObject)
     -- Construct the nameText based on specName and unitID settings
     if specName then
         if showSpecName and showArenaID and unitID then
-            nameText = specName .. " " .. unitID
+            local arenaNumber = string.match(unitID, "%d+")
+            nameText = specName .. " " .. (arenaNumber or "")
         elseif showSpecName then
             nameText = specName
         elseif showArenaID and unitID then
@@ -802,6 +827,9 @@ local function PlayerFrameNameChanges(frame)
         frame.bbfName:SetText("")
         return
     end
+    if not changeUnitFrameFont then
+        frame.bbfName:SetFont(frame.name:GetFont())
+    end
     if classColorTargetNames then
         ClassColorName(frame.bbfName, unit)
     end
@@ -826,6 +854,10 @@ local function TargetFrameNameChanges(frame)
     frame.name:SetAlpha(0)
     if not frame.unit then return end
     local unit = frame.unit
+
+    if not changeUnitFrameFont then
+        frame.bbfName:SetFont(frame.name:GetFont())
+    end
 
     if targetAndFocusArenaNames and IsActiveBattlefieldArena() then
         SetArenaNameUnitFrame(frame, unit, frame.bbfName)
@@ -880,6 +912,9 @@ local function PetFrameNameChanges(frame)
     if hidePetName then
         frame.bbfName:SetText("")
         return
+    end
+    if not changeUnitFrameFont then
+        frame.bbfName:SetFont(frame.name:GetFont())
     end
     frame.bbfName:SetText(frame.name:GetText())
     if classColorTargetNames then
@@ -937,6 +972,11 @@ local function TargetFrameToTNameChanges(frame)
     frame.name:SetAlpha(0)
     if not frame.unit then return end
     local unit = frame.unit
+
+    if not changeUnitFrameFont then
+        frame.bbfName:SetFont(frame.name:GetFont())
+    end
+
     if targetAndFocusArenaNames and IsActiveBattlefieldArena() then
         SetArenaNameUnitFrame(frame, unit, frame.bbfName)
     else